Refactor Channels, Presence and Messages (Yavor Georgiev)
There is a lot that can be shared between the Realtime and REST implementations of Channels, Presence and Messages, so this path mostly deals with getting things in place in anticipation of the sweet, sweet code reuse.

In the mean time this adds the `channels` attribute on `ARTRest`, as well as tests for channel creation, publishing and presence.