Replace ?-chains with force-unwraps.
There is some abuse of ?-chaining in the test code. If we expect a
thing to be there and if not being that the case would mean that
test code (e. g. a helper) is wrong, force-unwrap should be used
instead of silently ignoring this condition with a ?-chain so that
the error is caught from the test framework. The test framework
should only tell us about the code being tested, not about the
code that tests the code being tested.