Rename payload -> data, refactor message data encoding.
Well, this took a bit of time to get right. It started just
wanting to comply with the spec, and access message data just by
doing `message.data` instead of `message.payload.payload`. So
I wanted to remove just the amount of code necessary to do that,
but ended up opening a can of worms and didn't stop until it
kind of made sense.

It seems like the whole ARTPayload business was designed with a
use case which didn't really fit the spec at all, so it was then
hacked here and there to kind of work in some (not all) cases
by bypassing the design. It was also really overengineered, with
a whole class hierarchy thingy just for a few predefined cases.
I simplified the whole thing a lot and made it fit the spec.

I could've just set up a custom `message.data` getter method that
extracted the data from the ARTPayload instance to make it comply
with the spec, but well, this would've hit us down the road
anyway.