Fix PresenceMessage decoding test.
The test was expecting PresenceMessages _not_ to be decoded,
which is the wrong behavior.

Plus, DataEncoder has been refactored so that it doesn't take
out arguments but return a custom DataEncoderOutput instance.
Calling Objective-C functions with out arguments is a real pain
and really unsafe.

Encrypted test fixtures for presence messages are now working too.