Rewrite ARTEventEmitter per the spec and adjust Connection.
Previously, ARTConnection managed state subscriptions and
ARTEventEmitter was just a shell. Now the heavy-lifting of
passing events around is done by ARTEventEmitter, which is
generic and can be reused by others, and ARTConnection just uses
it to emit events and expose its listening interface.

This also introduces ARTConnectionStateChange, used in connection
state events' callbacks, as the spec says.

Something that went away is that starting listening to events
dispatched the current event to the callback. Now it doesn't. It
starts listening to _changes_ from that point on. I think this
makes more sense and is consistent with what you expect from an
event emitter.

Fixes #133.