Actually call {at,de}tach callbacks on success.
The callback was being subscribed to the message's ACK, while for
ATTACH and DETACH there's no ACK. Instead, an {AT,DE}TACHED
message is received. I've moved the callback call there, using
EventEmitter.

There was only a test for the attach timeout, which did work for
the subtle behavior that channel transitioning always cancels the
timeout. I've added tests for the happy paths as well.