Move IV argument to private Crypto constructor.
When decrypting, IV is being extracted from the beginning of the
message's data blob. When encrypting, the IV should be random, and
there's no point in taking it as an argument except for testing.

I've added and end-to-end test that sends and receives encrypted
messages. I've also run it with ably/ably-js as sender and receiver
and it works.