Serialize state transitions in a GCD queue.
State transitions can occur both from WebSockets (which run on their
on their own threads) and from user's actions, so they are intrinsically concurrent and need to be explicitly synchronized.

The library is still not thread-safe, but it's a start and a big
part of achieving at least internal thread-safety.