Sentry. (#603)
* WIP: Report crashes to Sentry.

To-do:

- [ ] Tests.
- [ ] Catch signals caused by the library (in progress).
- [ ] Send device info with report.
- [ ] Get stack trace line numbers to work in Sentry.

* Sentry: send device info.

* ARTChannels doesn't really have a ARTRest.

* Crash reports with (almost) full device info.

* Fix ARTRest exception handler on constructor.

* ARTRest: don't recover from exceptions in exception-recovering code.

* Don't recover from expected exceptions.

* Install KSCrash once, on ARTRest's constructor.

* ARTSentry: only report crashes from Ably.

* Wrap every callback from the user to call asynchronously.

This way, crashes happening in them won't be reported, since their
stack trace won't contain any Ably method.

* Report to Sentry on unexpected message serial.

* Use KSCrashAblyFork.

* Serialize state transitions in a GCD queue.

State transitions can occur both from WebSockets (which run on their
on their own threads) and from user's actions, so they are intrinsically concurrent and need to be explicitly synchronized.

The library is still not thread-safe, but it's a start and a big
part of achieving at least internal thread-safety.

* Bring back breadcrumbs.

* Bump KSCrashAblyFork version.

* Prevent deadlock when dispatch_syncing twice to same queue.

* Gzip Sentry reports.

* Serialize ARTLog.