Test Suite: fix [ARTPresenceMessage decodeWithEncoder:error:]: unrecognized selector sent to instance
Because of 8fb2e0c4616a57b7410e1cbcd2f446f905e7fd32, there are 2 ARTBaseMessage (ARTPresenceMessage) class method swizzling and Aspects doesn't support that and that's why the error occurs. There are similar issues in the Aspects repo like this one: Aspects/issues/145. I tried the fix of that PR but it didn't work. So, I decided to change all of the class methods swizzling to fix this case and reduce future issues.