Fix RTN10c: force the first connection to Failed state
The solution: "do not try and recover a connection from a client library instance that is still present and active."

Discussion: "the connection that you connected with at first is disconnected on line ~558 with 'transport superseded', because the second, new connection that you started recovered that connection.
But that first connection doesn't know that! All it knows is that it got a 'disconnected', and it should then try to connect again.
So it then tries that -- and steals the connection back from the second. (this time getting a 'Connection superseded' for the reason I said above, because it had an old handle).
So you have these two library instances fighting over the same connection, trading it back and forth."