[WIP] Replace weak refs by explicit cycle breaking where possible.
Throughout the codebase, we were liberally using weak references to
break all possible ref cycles, even those that are temporary and should
be broken explicitly. Other weak refs are there just for no good reason,
just to signify that we don't expect an object to lose all other
references to it before our reference is lost.

This by itself results in brittle code. The default should be strong,
which ensures a reference won't disappear before you expected. In other
cases, a more explicit handling of the reference lifecycle (e. g.
explicitly calling a "close" or "cancel" method) makes things clearer.

This PR removes all weak references except those where there's a stable
cycle and the ownership relationship between the objects in the cycle
is clear; for example, ARTRealtime clearly "owns" ARTRest, insofar their
lifetimes are entangled: ARTRealtime initializes ARTRest, and when the
ARTRealtime is released, so should be its ARTRest, which should have no
other references to it. All those cases have been documented.

In some places, we relied on weak references to implicitly invalidate
timers, enqueued operations and things like that. The pattern looks like
this:

    __weak MyClass *weakSelf = self;
    scheduleBlockToBeCalledLater(^{
        [weakSelf someMethod];
    });

The intent here is to cancel scheduled operations on the object if we
stop being interested in the object. And this "interest" is expressed
as holding a reference to the object somewhere.

This way we're forcing the user to hold on to their reference to (say)
ARTRealtime even if they're not going to use it anymore. That's not
what one would expect in a garbage-collected language, in which other
client libraries are written. In those, you would explicitly call
realtime.close if you want it to stop operating.

So this change makes this behavior uniform with other client libraries,
but now we're forced to explicitly unschedule the operations on close.

For this, a little helper class called ARTFlusher to register things
that should be "flushed" (cancelled, emptied, etc.) once we call "close"
or similar on the object they operate on. We flush this way event
listeners, timers, enqueued messages, etc. related to a realtime or
realtime channel instance.