Split ARTRealtime in public and internal objects.
The public object is just a proxy for the internal object. The purpose
of the split is to avoid any thread/queue other than Ably's internal
queue to deallocate the internal object. This was causing a concurrency
bug in which some Ably code uses a weak reference to the object while
the last strong reference is released in the user's thread, causing a
concurrent deallocation and a crash.

Now, user's code can only reference and deallocate the public object.
If that public object is deallocated, the trick is to create a new
reference to the internal object, and schedule the releasing of such
reference in Ably's queue. This ensures that uses of weak references to
this internal object isn't concurrent with its deallocation, as long as
no references to the internal object outliving the public object's
lifetime end up in some other thread/queue.

This is way too complicated but unfortunately necessary if the library
is to be safely accessed by any thread, to which we're bound by the
semver compatibility guarantees.

ARTRealtime is only one of the objects that need to be split this way;
every object to which user code may hold a reference needs it too. The
process is pretty mechanic but rather tedious:

- Extract ARTFooProtocol with ARTFoo's public API.
- Make ARTFoo implement ARTFooProtocol; properties should have just
  getters.
- In ARTFoo+Private.h, extend ARTFoo and add:

    @interface ARTFoo ()

    - (void)internalAsync:(void (^)(ARTFooInternal *))use;

    @end

- In ARTFoo+Private.h, declare ARTFooInternal's interface, synthesizing
  properties as needed.
- Make all previous private extensions on ARTFoo be on ARTFooInternal
  instead.
- Declare a private queue property on ARTFooInternal so that ARTFoo can
  access it. 
- On ARTFoo.m, replace ARTFoo with ARTFooInternal.
- Implement ARTFoo like this:

    @implementation ARTFoo {
        ARTFooInternal *_internal;
    }

    - (void)dealloc {
        __block ARTFooInternal *internal = _internal;
        dispatch_async(_internal.queue, ^{
            internal = nil;
        });
    }

    - (void)internalAsync:(void (^)(ARTFooInternal * _Nonnull))use {
        dispatch_async(_internal.queue, ^{
            use(self->_internal);
        });
    }


- Create initializers of ARTFoo that initialize and set to _internal
  a ARTFooInternal. For example:

    - (instancetype)initWithOptions:(ARTClientOptions *)options {
        self = [super init];
        if (self) {
            _internal = [[ARTRealtimeInternal alloc] initWithOptions:options];
        }
        return self;
    }

- Redirect the rest of the methods of ARTFoo to _internal. For example:

    - (void)time:(void (^)(NSDate *_Nullable, NSError *_Nullable))cb {
        [_internal time:cb];
    }

- Replace usages of ARTFoo throughout the codebase that need to access
  ARTFooInternal's private methods by ARTFooInternal. In tests too.
- Double-check that no reference to ARTFooInternal can be kept by user's
  code, ever.