Extract queued deallocator to shared object amongst public objects.
Instead of repeating the same pattern in all public objects, do it once
in a single object that always points to RealtimeInternal (and, later,
RestInternal) which schedules its deallocation.

The async dealloc only needs to happen for RealtimeInternal; each public
object's internal counterpart can be deallocated normally, since there
will be a reference path from RealtimeInternal to that internal object
anyway, so the object will only be deallocated when RealtimeInternal is
deallocated anyway.

The picture now looks like this:

```
                                           
    +----------+     +----------------+    
    | Realtime |---->|RealtimeInternal|    
    +--+-------+     +-^------+-------+    
       |               |      |   ^        
       v             +-+      |   |        
     +-------------+ |        |   |        
     |QueuedDealloc| |        |   |        
     +-------------+          |   |        
       ^                      |   |        
       |                      v   |        
    +--+-------+       +----------+-------+
    |Connection|------>|ConnectionInternal|
    +----------+       +------------------+
```