Add *Channels.iterate method.
This unfortunately breaks the API, but it follows the spec, and there's
no easy way around it now that ART*Channels has an internal counterpart
that actually holds the references while ART*Channels creates public
wrappers for them on demand. You can't implement NSFastEnumeration over
objects that don't outlive it.