Copy channels for public iterate() on internal queue. (#919)
We were getting the iterator in the internal queue, but then doing the
copy into public wrappers in the caller's queue. So the internal queue
could then concurrently run a task that mutates the collection while
it's being copied.

Now the copying is done while still on the internal queue, on the common
ARTChannels; Rest and Realtime just provide mappers to copy into the
right public channel type.