Set logLevel=verbose every time setupOptions is called with debug=true. (#927)
We were only setting it when testApplication was nil, ie. on the first
test or if forceNewApp was also true. But that option has nothing to
do with the testApplication, it applies to the ClientOptions instance
produced, whether it's for a fresh Ably application or not.

This allows you to see verbose logs for any given test, not just the
first one.