Don't have push state machine singleton.
Instead, just cache it in the Rest instance.

Otherwise, if the user stops using the first instance and activates push
through a new instance, push activation will still use the first
instance an things won't work as expected.

Unfortunately, this now means that it's possible to have more than one
state machines per process, which is bad; RSH3 mandates that the state
machine must be a single app-wide concurrent process.

Once #966 is fixed, this drawback may be lifted. For now, is has been
documented in the README.