De-abstract shouldRenewToken, fix connection establishment logic.
Reconnection after DISCONNECTED and failing to establish a connection
due to token error follow similar but slightly different logics. It's
not worth trying to abstract it.

Now, per RTN14b, of connection establishment failure: "If  the attempt
to create a new token fails, or the subsequent connection attempt fails
due to another token error, then the connection will transition to the
DISCONNECTED state, and the Connection#errorReason should be set. (If no
means to renew the token is provided, RSA4a applies [ie. moves to
FAILED])"