Fix handling of token error when first connecting. (#986)
* Document why we check time offset before checking token expiry.

* Add a test for RSA4b happy path on REST request.

* Fix handling of token error when first connecting.

Removed complex and apparently unnecessary logic on error protocol
message handler. No tests broken by this.

* RSA4c1&2 test: Don't check errorReason until settled on DISCONNECTED.

* Don't force-unwrap stateChange.reason in test assertions.

Soft-unwrapping catches assertion failures more gracefully.

* De-abstract shouldRenewToken, fix connection establishment logic.

Reconnection after DISCONNECTED and failing to establish a connection
due to token error follow similar but slightly different logics. It's
not worth trying to abstract it.

Now, per RTN14b, of connection establishment failure: "If  the attempt
to create a new token fails, or the subsequent connection attempt fails
due to another token error, then the connection will transition to the
DISCONNECTED state, and the Connection#errorReason should be set. (If no
means to renew the token is provided, RSA4a applies [ie. moves to
FAILED])"

* Make test for RTN15h1 more precise.

Test that we actually went through CONNECTED, instead of going directly
to FAILED (that would be RTN14b).

Also, don't assert on the error message. Asserting on the code is
enough.

* Fix test for RTN15h2: "will transition to the DISCONNECTED"

... not FAILED.

Also, make it faster.