Compile SocketRocket directly into the SDK
We remove it as a dependency in Carthage and CocoaPods, and deal with
the fallout. There is some further work needed to fix the CocoaPod,
which I’ll do in a separate commit.

For future reference, we are doing this as part of a wider effort to
reduce the number of transitive dependencies that we impose on users of
the SDK. See issue #949 for more details.

I manually removed the framework copy steps. I don’t _think_ there is a
more official Carthage-provided way of doing this, but I could be quite
wrong; I don’t know much about Carthage.