Remove instance variables from PushAdmin tests
There’s no need for these to be instance variables. We’re not mutating
anything, it’s just static testing data.

The motivation for this change is that it’s groundwork for removing the
Quick testing framework using an automated migrator tool (#1201). The
migrator that I’m writing won’t be able to handle the capture lists in
the QuickSpec method call trailing closures, so I’m replacing them.