Unroll a test-defining loop in Crypto tests
The existing implementation used a loop to implement a form of
table-based testing (i.e. define a list of tests at runtime). Here I
remove this loop, extracting its body to a function (whose name begins
with `reusableTests`) and replacing the iteration with a static list of
invocations fo this function.

The motivation for this change is that it’s groundwork for removing the
Quick testing framework using an automated migrator tool (#1201). The
migrator that I’m writing only knows how to handle a specific set of
statements at a `context`/`describe` level, and would not know what to
do with e.g. an expression like a `forEach` call. (It wouldn’t be able
solve this problem generally without executing the test code during the
migration.)

Hence we replace these expressions with something that the migrator
knows how to handle. The `reusableTests` prefix is important here — the
migrator _does_ know how handle a call to this kind of function and
knows that, unlike other function declarations, it should look inside
the body of this function, where it will find more tests that it needs
to migrate.