Unroll a test-defining loop in RealtimeClientChannel tests
Same motivation as 30f8448. The solution is simpler here, since the loop
only contains a single call to `it`, so after we pull the loop body out
to a reusable function, that function no longer contains any Quick
method calls and hence the migrator doesn’t need to look inside its
body.