Remove a floating statement from RealtimeClient tests
The motivation for this change is that it’s groundwork for removing the
Quick testing framework using an automated migrator tool (#1201). The
migrator that I’m writing won’t know what to do with arbitrary
statements at a `spec` / `describe` / `context` level. It does, however,
know what to do with variable declarations. So, let’s manually move this
statement inside the declaration of the variable that it’s related to.