Fix setting of fallback retry timeout in tests
I’m pretty sure that the existing code was unintentional. In Quick, code
at the top level of `context` is executed as part of the initial setup
of the entire test suite, before any test cases are executed. This means
that the effect of the existing code is to use a fallback retry timeout
of 10 seconds for the whole test suite, until it is changed to 1 second
during the execution of the test for RSC15e.

I believe that the intention of the existing code that modifies the
timeout was to make sure that the specified timeout is in place during
the execution of the test cases within that particular context, so I’ve
changed the behaviour to do this.

This marks a behavioural change to the tests but I think it’s in line
with the original intentions.

Despite the existing code having this issue, I believe that the affected
tests were still passing for the following reasons:

- "should restore default primary host after fallbackRetryTimeout
  expired": this test needs a timeout of 1 second, but coincidentally
  the previous test for RSC15e has already set the timeout to 1 second;

- "should use another fallback host if previous fallback request failed
  and store it as default if current fallback request succseeded": this
  test is trying to specify a timeout of 10 seconds, but in fact when this
  test executes the timeout is set to 1 second. However, I think the idea
  of the 10 seconds was just to make sure that the client doesn’t stop
  using fallback hosts during the execution of the test, and 1 second is
  still plenty of time for the test to execute, so it doesn’t make
  a difference.

The original motivation for this change was the same as eb1ea54 —
removing arbitrary statements at `context` level.