Move two statements at top level of `context` into `beforeEach`
I’m pretty sure that the existing code was an unintentional mistake.
Code at the top level of `context` is executed as part of the initial
setup of the entire test suite, before any test cases are executed. This
means that the effect of the existing code is to use a fallback retry
timeout of 10 seconds for the whole test suite. I believe that the
intention of the code that changes the time was to limit it to the test
cases within that particular context, so that’s what I’ve done. This
marks a behavioural change to the tests but I think it’s in line with
the original intentions.