Replace call to `pending` with `xit`
The motivation for this change is that it’s groundwork for removing the
Quick testing framework using an automated migrator tool (#1201). The
migrator that I’m writing won’t know what to do with the body of the
closure passed to `pending` — since Quick doesn’t execute this closure,
its contents are a black box to us.