Get rid of a class defined at a `describe` level
Like 2f047e1, not sure why we need a class to hold a variable that’s
used in a single test case, but nor do I fully understand what this test
is doing…

Like 252e887, this is a behaviour-modifying change, since it delays the
moment initialization of the AblyClient instance. Not sure whether this
matters here.