Reduce number of local variables in Stats tests
These variables are actually only used to set up some other variables,
so we move them inside those variables’ declarations.

The motivation for this change is that it’s groundwork for removing the
Quick testing framework using an automated migrator tool (#1201). The
migrator that I’m writing will pull all `spec` / `describe` / `context`
level local variables to a single namespace for the whole file, so we
need to avoid clashes. The easiest way in this particular case is to
just remove the variables.