Remove a few unnecessary context-level variables in Stats tests
I happened to notice that these variables are only used in setting up
other variables, so let’s move them inside the declaration of those
variables.

The motivation for this change is that it’s (not strictly necessary)
groundwork for removing the Quick testing framework using an automated
migrator tool (#1201). The migrator that I’m writing will pull all
context-level variable declarations in a file into a single, potentially
large, flattened list of declarations. So it would be nice to make this
list as small as can be. (I'm not saying that the list _is_ now "as
small as can be", I'm just fixing a few things that I happened to
notice.)