Rename a `describe`-level variable called `name`
The motivation for this change is that it’s groundwork for removing the
Quick testing framework using an automated migrator tool (#1201). The
migrator that I’m writing will pull all `spec` / `describe` / `context`
level local variables outside of the class declaration and into
fileprivate global variables. But if we let the migrator hoist this
particular variable to a global with the same name, then it’ll be
shadowed by XCTestCase’s `name` property, which we don’t want.