Rewrite test descriptions containing meaningful symbols
The motivation for this change is that it’s groundwork for removing the
Quick testing framework using an automated migrator tool (#1201). The
migrator that I’m writing will convert all symbols in test descriptions
to underscores in the emitted method names. This is okay most of the
time, but we have a few places where the symbols have meaning, and
removing them will change the description of the test. The migrator
emits a warning when it thinks this is the case, but it’s up to us to
look at this output and decide which ones need changing.