Suppress compiler warning for deprecated property
Our tests are using the deprecated
-ARTClientOptions.fallbackHostsUseDefault property. We can’t remove this
usage until we remove the property. But we want to be able to turn on
“treat warnings as errors” so we need a way to suppress the resultant
compiler warning.

Swift doesn’t really provide options for suppressing warnings, so
instead we take advantage of the fact that it doesn’t emit warnings when
using deprecated functionality inside of a method which is itself marked
as deprecated. See
https://forums.swift.org/t/swift-should-allow-for-suppression-of-warnings-especially-those-that-come-from-objective-c/19216/68.