Don’t copy argument in testSuite_getArgumentFrom:
We want to handle arguments that don’t support NSCopying.

I assume the original intention was to return something to the caller
that couldn’t be mutated under their feet, but I think I’d expect a spy
method like this to accurately report the exact instance that was
received.