When setting realtime channel options, pass through to REST channel
When calling -[ARTRealtimeChannel setOptions:callback:], the passed
options are not used to update the options of this instance’s REST
channel. This means that, for example, if we update the realtime
channel’s cipher options, these options will not be used to decrypt
messages subsequently fetched using the -history:* methods.

I think this is something we missed when fixing #1207 (34d68ec).

This issue is causing ably/ably-flutter#296.

Closes #1265.