Emit error information from ARTDeviceStorage loading methods
We’re not doing anything with this information yet. Also, the
NSUserDefaults implementation of objectForKey: never returns an error,
although data protection policies may mean that it’s not possible to
fetch a value from user defaults. We need to figure out how to detect
this situation.

The change to the method signature (passing the return value indirectly
through a reference) is due to the following Swift compiler error:

> Throwing method cannot be an implementation of an @objc requirement
> because it returns a value of optional type 'Any?'; 'nil' indicates
> failure to Objective-C