Update Fastlane
Gonna be making some changes to its config so seems like a good time to
make sure it’s up to date.

update fastlane again

has some changes to handling skipped tests

i'm a bit confused about what's happening with the trainer repo - is
that still being worked on or is it now part of fastlane? because e.g.
the skipped test changes aren't there afaik
https://github.com/fastlane/fastlane/pull/19957/files

we need to get this in the dev instructions too