Only try to test on versions of iOS / tvOS that are available to us
Our GitHub runners’ virtual environment only makes a limited set [1] of
iOS and tvOS runtimes available for each Xcode version. We currently
build the test suite using the default (i.e. latest) version of Xcode
available on the runner, and I believe that we don’t intend to change
this.

So, we need to only attempt to run against the runtimes that are
actually available. This means only testing against later iOS versions
(and hence removing one of our workflows).

[1] https://github.com/actions/virtual-environments/blob/754215539971e726fa6989689246a2da00544c57/images/macos/macos-10.15-Readme.md