Add a very primitive way of running tests for 1 day straight
Can look into some way of de-duping this maybe (e.g. generate with a
script). And in theory we can go up to 35 days (workflow limit).

I tried doing this instead using the `job.id.concurrency` option - but
some jobs just seemed to fail immediately instead of waiting for the
other job holding the lock to complete, so gave up on that without
investigating much more.