Upload test crash reports to observability server
The test failure message that we get when a test crashes (e.g. because
of an exception) is not very informative; it’s usually something like

Crash: xctest (61641): Namespace SIGNAL, Code 0x4 (#CharacterRangeLen=0)

This tells us very little about the reason for the crash and makes it
difficult to investigate. So, here I’m extracting the crash reports
(.crash files) from the .xcresult bundle, and uploading them to the
observability server alongside the JUnit report. This will give us much
more information about the reason for the crash – which function and
thread it occurred in, the exception message, etc.

We’re using the xcparse tool, which is able to extract attachments (such
as crash reports) from .xcresult bundles. We’re using an Ably fork of
this tool which also outputs a JSON file describing the test cases that
attachments belong to.