Fix test case RealtimeClientPresenceTests.test__037__Presence__update__should_update_the_data_for_the_present_member_with_a_value
This fixes the error

> Waited more than 20.0 seconds
> (/Users/runner/work/ably-cocoa/ably-cocoa/Spec/Tests/RealtimeClientPresenceTests.swift#CharacterRangeLen=0&EndingLineNumber=1183&StartingLineNumber=1183)

This happens because, depending on timing, the ENTER presence action
that we publish (more or less simultaneously with attaching to the
channel) will sometimes be reflected by Realtime in the channel’s
initial post-attach SYNC.

In this case, the SDK will ignore any subsequent matching ENTER messages
that it receives and will not send them to subscribers, causing the
aforementioned timeout.

So, we avoid this scenario by waiting for any post-attach SYNC to
complete before publishing the ENTER presence action.

Test case:
https://test-observability.herokuapp.com/repos/ably/ably-cocoa/test_cases/114e01a8-8971-4c3a-9cb7-0bb128e2749e

Example failure:
https://test-observability.herokuapp.com/repos/ably/ably-cocoa/failures/48dce95a-dca8-4b3c-952d-865277d63b73

Closes #1309.