Take more care to catch received SYNC message in test
In the current setup of this test, we are hoping that the callback to
transport.setListenerBeforeProcessingIncomingMessage gets called before
the incoming SYNC message gets processed. But since this callback gets
called asynchronously on the main thread, it seems possible to me that
this might not happen.

So, instead of trying to set up code that, if executed in time, will
ignore the SYNC message when it comes in, let’s ignore it _whilst_
processing the incoming messages.

This is just speculation on my part, I haven’t directly observed the
described problem happening. But we’re trying to fix an intermittent
failure in this test and it seems like a plausible issue I think it’s
worth addressing.