Fix skipped RealtimeClientPresenceTests.test__009
And re-enable it.

This test made a (I believe previously valid) assumption that Realtime
would not set the HAS_PRESENCE flag on an ATTACH protocol message for a
channel that has no members present. This assumption became invalid with
a Realtime change that made it possible for this to happen (as described
in the Slack thread you’ll find in the diff here).

I didn’t know it at the time, but looks like this Realtime change is
also responsible for the failures described and fixed in 7c53cec,
57b7266 and 3f155f9.