Build Carthage dependencies in all architectures when running `make update`
This stops the usage of
Scripts/carthage-with-workaround-for-issue-3019.sh when building the
project’s dependencies using Carthage for local development (i.e. when
running `make update` or one of its sub-commands).

This file was seems to have been originally introduced to fix an error
related to duplicate architectures when running Carthage. I don’t get
any errors switching back to plain Carthage without a wrapper when
running `make update` on my M1 Pro machine, using Carthage 0.38.0. Given
that the linked issue https://github.com/Carthage/Carthage/issues/3019
is now marked as resolved, I guess something was changed in Carthage
that resolves this for us 🤷.

This closes #1506, where we were seeing "Undefined symbol:
_OBJC_CLASS_$_ARTDeltaCodec" errors running `make test_iOS` on M1*
machines. I guess in the part of the script where we exclude ARM
architectures (I don’t exactly know what that means) we had stopped
building something that we needed to run on the current platform.

We’re still using this file for the `make carthage_package` task. We
should check whether we need it there too, but since that might have
user-facing consequences (instead of just breaking our development
environment) I don’t want to do that hastily. I’ve created a separate
issue #1509 for that.