Lock to using Xcode 13.4.1 in check-pod CI job
This job recently started using Xcode 14 by default (this sort of thing
can happen since we’re using the macos-latest image).

This fixes build scripts that use
Scripts/carthage-with-workaround-for-issue-3019.sh, which we haven’t yet
updated for Xcode 14. I’d prefer not to spend time updating that script
now since hopefully we can get rid of it soon (see #1510, #1509).