Remove protocolMessagesLogger from ARTWebSocketTransport
This was introduced in eb3a8f6, where as far as I can tell it did
nothing useful, since the logger (which is using ARTLog’s default
behaviour) only logs messages of level `warn` or above, and we only ever
log to it at level `info`, hence it never logs anything.

I’m not sure what the original intention was behind this logger, but it
seems safe to remove it.

Closes #1556.