Let CocoaPods generate its own module map and umbrella header
This means we don’t need our CocoaPods-specific module map, so one less
file to maintain. This means that users of our pod can no longer import
the Ably.Private module, but they shouldn’t be doing that anyway; I
believe that just exists for use in the library’s unit tests.

TODO oh the tests use Ably.Private too, as does SoakTests

I do however have a feeling that we might be exporting what we think is
Ably.Private even as part of just Ably module — if I remove the imports
to Ably.Private in the tests I can still access stuff in it

https://guides.cocoapods.org/syntax/podspec.html#module_map