Don’t expose ARTHTTPExecutor.logger property
This is part of #1644. The aim of this issue is for every object to log
to its own logger, instead of grabbing a logger from some other object
it has access to. This will allow us to subsequently embed contextual
information in the logger such as the identity of the object it belongs
to.